Don’t set id in insert! in tests or seeds
If ids are set explicitly, Postgres does not increment the next value of
the primary key sequence. This lead to Postgres trying to insert new
values with an automatically incremented value for id that has already
been inserted explicitly. In that case, the test failed with a unique
constraint violation.

Adapt Api.Helpers.with_authorization to take a User as parameter instead
of hardcoding the account as module variable.