Remove dependency on phoenix_html
phoenix_html, in the previous setup, was expected to be located at
`../deps/phoenix_html`. That required `mix deps.get` to have been run
because it was configured as a local dependency, not loaded from npm.
Currently, we don’t want to setup Elixir when we test Elm code because
that would defeat the purpose of having a build matrix.