Update directory structure
The structure now matches the expectations of phoenix >= 1.3.

The dependency on phoenix itself has already been > 1.4 for quite some
time, but when it was updated to 1.3 the changes to the directory
structure had not been made.

The changes have been made now because phoenix has updated its
generators to make use of the new structure, and there are plans to use
`phx.gen.live` in the near future.