Increase length of title and url for entry
I found a feed that hat URLs larger than 256 characters which let the
application crash with a SQL exception. This is a quick fix that will
prevent the application from crashing in this specific case, but it does
not yet deal with the larger issue of proper error handling and not
crashing in the first place.