Add ExRss.Application
This replaces `ExRss`. This change was introduced in Phoenix 1.3, but I
hadn’t done the migration yet.

This has the side-effect of addressing deprecation warnings related
to applications not being started because they were not defined in
`applications` or `extra_applications`. With this change, applications
from dependencies are automatically picked up with no extra
configuration needed.